CheckBoard, ScoreBoard 분리, 상속 추가
